SCP-1921: SigningProcess: tidying up
- use explicit exports
- move declarations to types module
- use newtype wrappers
