+++ plutus-pab/src/Cardano/SigningProcess/Types.hs
@@ -7,15 +7,36 @@
 {-# LANGUAGE TypeApplications  #-}
 
 module Cardano.SigningProcess.Types
-    ( SigningProcessMsg (..)
+    ( SigningProcessEffects
+    , SigningProcessMsg (..)
+    , SigningProcessConfig (..)
+    , WalletUrl
     ) where
 
+import           Control.Monad.Freer.Error     (Error)
 import           Data.Aeson                    (FromJSON, ToJSON)
 import           Data.Text.Prettyprint.Doc     (Pretty (..), (<+>))
 import           GHC.Generics                  (Generic)
 
 import           Cardano.BM.Data.Tracer        (ToObject (..))
 import           Cardano.BM.Data.Tracer.Extras (Tagged (..), mkObjectStr)
+import           Cardano.Wallet.Types          (WalletUrl)
+import           Control.Monad.Freer.State     (State)
+import           Wallet                        (SigningProcessEffect)
+import qualified Wallet.API                    as WAPI
+import           Wallet.Emulator.Wallet        (SigningProcess, Wallet)
+
+type SigningProcessEffects =
+    '[ SigningProcessEffect, State SigningProcess, Error WAPI.WalletAPIError]
+
+
+data SigningProcessConfig =
+    SigningProcessConfig
+        { spWallet  :: Wallet -- Wallet with whose private key transactions should be signed.
+        , spBaseUrl :: WalletUrl
+        } deriving stock (Eq, Show, Generic)
+          deriving anyclass (ToJSON, FromJSON)
+
 
 newtype SigningProcessMsg =
     StartingSigningProcess Int -- ^ Starting up on the specified port
