+++ plutus-pab/src/Plutus/PAB/Types.hs
@@ -15,8 +15,8 @@ import           Cardano.BM.Data.Tracer.Extras  (StructuredLog (..))
 import qualified Cardano.ChainIndex.Types       as ChainIndex
 import qualified Cardano.Metadata.Types         as Metadata
 import           Cardano.Node.Types             (MockServerConfig (..))
-import qualified Cardano.SigningProcess.Server  as SigningProcess
-import qualified Cardano.Wallet.Types           as WalletTypes
+import qualified Cardano.SigningProcess.Types   as SigningProcess
+import qualified Cardano.Wallet.Types           as Wallet
 import           Control.Lens.TH                (makePrisms)
 import           Data.Aeson                     (FromJSON, ToJSON (..))
 import qualified Data.HashMap.Strict            as HM
@@ -100,7 +100,7 @@ data DbConfig =
 data Config =
     Config
         { dbConfig                :: DbConfig
-        , walletServerConfig      :: WalletTypes.WalletConfig
+        , walletServerConfig      :: Wallet.WalletConfig
         , nodeServerConfig        :: MockServerConfig
         , metadataServerConfig    :: Metadata.MetadataConfig
         , pabWebserverConfig      :: WebserverConfig
